backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
pymongo==4.4.1
python-dotenv==1.0.0
bcrypt==4.0.1
transformers==4.30.2
torch==2.0.1
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
requests==2.31.0
